+++ zarrs_metadata/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `GroupMetadataV2` constructors
 - Add `ArrayMetadataV2` constructors
+- Implement `From<{&str,String}>` for `DataTypeMetadataV2`
 
 ### Changed
 - **Breaking**: Mark `GroupMetadataV3` and `ArrayMetadataV3` as non-exhaustive